+++ android/anc/src/main/java/org/smartregister/fhircore/anc/ui/report/ReportHomeActivity.kt
@@ -43,12 +43,20 @@ import java.io.InputStream
 import java.util.concurrent.Executors
 import org.hl7.fhir.instance.model.api.IBaseBundle
 import org.hl7.fhir.instance.model.api.IBaseResource
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
+import androidx.core.util.Pair
+import com.google.android.material.datepicker.MaterialDatePicker
+import java.util.Calendar
 import org.smartregister.fhircore.anc.AncApplication
 import org.smartregister.fhircore.anc.R
 import org.smartregister.fhircore.anc.data.report.ReportRepository
 import org.smartregister.fhircore.engine.cql.LibraryEvaluator
 import org.smartregister.fhircore.engine.cql.MeasureEvaluator
 import org.smartregister.fhircore.engine.data.remote.fhir.resource.FhirResourceDataSource
+import org.smartregister.fhircore.anc.ui.report.ReportViewModel.ReportScreen
 import org.smartregister.fhircore.engine.ui.base.BaseMultiLanguageActivity
 import org.smartregister.fhircore.engine.ui.questionnaire.QuestionnaireActivity
 import org.smartregister.fhircore.engine.ui.theme.AppTheme
@@ -119,6 +127,9 @@ class ReportHomeActivity : BaseMultiLanguageActivity() {
     val dispatcher: DispatcherProvider = DefaultDispatcherProvider
     reportViewModel = ReportViewModel(repository, dispatcher)
     reportViewModel.backPress.observe(
+    val repository = ReportRepository((application as AncApplication).fhirEngine, patientId, this)
+    val viewModel = ReportViewModel.get(this, application as AncApplication, repository)
+    viewModel.backPress.observe(
       this,
       {
         if (it) {
@@ -346,5 +357,38 @@ class ReportHomeActivity : BaseMultiLanguageActivity() {
 
   fun handleCQLMeasureLoadPatient(auxPatientData: String) {
     patientDetailsData = auxPatientData
+    viewModel.showDatePicker.observe(
+      this,
+      {
+        if (it) {
+          showDateRangePicker(viewModel::onDateSelected)
+        }
+      }
+    )
+
+    setContent {
+      val reportState = remember { viewModel.reportState }
+      AppTheme { ReportView(viewModel) }
+    }
+  }
+
+  @Composable
+  fun ReportView(viewModel: ReportViewModel) {
+    // Choose which screen to show based on the value in the ReportScreen from ReportState
+    when (viewModel.reportState.currentScreen) {
+      ReportScreen.HOME -> ReportHomeScreen(viewModel)
+      ReportScreen.FILTER -> ReportFilterScreen(viewModel)
+      ReportScreen.PICK_PATIENT -> ReportFilterScreen(viewModel)
+      ReportScreen.RESULT -> ReportResultScreen(viewModel)
+    }
+  }
+
+  private fun showDateRangePicker(onDateSelected: (Pair<Long, Long>?) -> Unit) {
+    val builder = MaterialDatePicker.Builder.dateRangePicker()
+    val now = Calendar.getInstance()
+    builder.setSelection(Pair(now.timeInMillis, now.timeInMillis))
+    val dateRangePicker = builder.build()
+    dateRangePicker.show(supportFragmentManager, dateRangePicker.toString())
+    dateRangePicker.addOnPositiveButtonClickListener { onDateSelected(dateRangePicker.selection) }
   }
 }