+++ android/anc/src/main/java/org/smartregister/fhircore/anc/ui/report/ReportViewModel.kt
@@ -17,6 +17,12 @@
 package org.smartregister.fhircore.anc.ui.report
 
 import androidx.lifecycle.LiveData
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.setValue
+import androidx.core.util.Pair
+import androidx.lifecycle.AndroidViewModel
+import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
@@ -24,6 +30,9 @@ import androidx.paging.Pager
 import androidx.paging.PagingConfig
 import androidx.paging.PagingData
 import ca.uhn.fhir.parser.IParser
+import java.text.SimpleDateFormat
+import java.util.Date
+import java.util.Locale
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.launch
 import org.smartregister.fhircore.anc.data.report.ReportRepository
@@ -38,9 +47,43 @@ class ReportViewModel(
 ) : ViewModel() {
 
   val backPress: MutableLiveData<Boolean> = MutableLiveData(false)
+  val showDatePicker: MutableLiveData<Boolean> = MutableLiveData(false)
+  val selectedMeasureReportItem: MutableLiveData<ReportItem> = MutableLiveData(null)
+  val simpleDateFormatPattern = "d MMM, yyyy"
+
+  private val _startDate = MutableLiveData("start date")
+  val startDate: LiveData<String>
+    get() = _startDate
+
+  private val _endDate = MutableLiveData("end date")
+  val endDate: LiveData<String>
+    get() = _endDate
+
+  private val _patientSelectionType = MutableLiveData("All")
+  val patientSelectionType: LiveData<String>
+    get() = _patientSelectionType
+
+  private val _isReadyToGenerateReport = MutableLiveData(true)
+  val isReadyToGenerateReport: LiveData<Boolean>
+    get() = _isReadyToGenerateReport
+
+  var reportState: ReportState = ReportState()
 
   fun getReportsTypeList(): Flow<PagingData<ReportItem>> {
     return Pager(PagingConfig(pageSize = PaginationUtil.DEFAULT_PAGE_SIZE)) { repository }.flow
+  }
+
+  fun onReportMeasureItemClicked(item: ReportItem) {
+    selectedMeasureReportItem.value = item
+    reportState.currentScreen = ReportScreen.FILTER
+  }
+
+  fun getSelectedReport(): ReportItem? {
+    return selectedMeasureReportItem.value
+  }
+
+  fun getPatientSelectionType(): String? {
+    return patientSelectionType.value
   }
 
   fun onBackPress() {
@@ -57,6 +100,53 @@ class ReportViewModel(
       val auxCQLLibraryData =
         parser.encodeResourceToString(fhirResourceDataSource.loadData(libraryURL).entry[0].resource)
       libraryData.postValue(auxCQLLibraryData)
+  fun onBackPressFromFilter() {
+    reportState.currentScreen = ReportScreen.HOME
+  }
+
+  fun onBackPressFromResult() {
+    reportState.currentScreen = ReportScreen.FILTER
+  }
+
+  fun onDateRangePress() {
+    showDatePicker.value = true
+  }
+
+  fun onPatientSelectionTypeChanged(newType: String) {
+    _patientSelectionType.value = newType
+  }
+
+  fun onGenerateReportPress() {
+    reportState.currentScreen = ReportScreen.RESULT
+  }
+
+  fun onDateSelected(selection: Pair<Long, Long>?) {
+    showDatePicker.value = false
+    if (selection == null) {
+      _isReadyToGenerateReport.value = false
+      return
+    }
+    val startDate = Date().apply { time = selection.first }
+    val endDate = Date().apply { time = selection.second }
+    val formattedStartDate =
+      SimpleDateFormat(simpleDateFormatPattern, Locale.getDefault()).format(startDate)
+    val formattedEndDate =
+      SimpleDateFormat(simpleDateFormatPattern, Locale.getDefault()).format(endDate)
+
+    _startDate.value = formattedStartDate
+    _endDate.value = formattedEndDate
+    _isReadyToGenerateReport.value = true
+    reportState.currentScreen = ReportScreen.FILTER
+  }
+
+  companion object {
+    fun get(
+      owner: ViewModelStoreOwner,
+      application: AncApplication,
+      repository: ReportRepository
+    ): ReportViewModel {
+      return ViewModelProvider(owner, ReportViewModel(application, repository).createFactory())[
+        ReportViewModel::class.java]
     }
     return libraryData
   }
@@ -144,5 +234,21 @@ class ReportViewModel(
       valueSetData.postValue(fullResourceString.toString())
     }
     return valueSetData
+  }
+
+  class ReportState {
+    var currentScreen by mutableStateOf(ReportScreen.HOME)
+  }
+
+  enum class ReportScreen {
+    HOME,
+    FILTER,
+    PICK_PATIENT,
+    RESULT
+  }
+
+  object PatientSelectionType {
+    const val ALL = "All"
+    const val INDIVIDUAL = "Individual"
   }
 }